feat: package enhanceEditor
